api/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-4-jet @ git+https://github.com/Mojtaba-saf/django-jet.git@c4b80834a12269ccffb4b85c1401d3b0a8b50817
django-crontab==0.7.1
djangorestframework==3.14.0
idna==3.4
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.12
